Moved matching code to a new subpackage of the 'internal' package
